Fix TranslationContext import paths after moving to types.ts
Updated import statements in:
- app/api/translate/route.ts
- lib/hooks/use-translation.ts
- lib/translation-batcher.ts

Changed from importing from llm-translate-client.ts to types.ts
to match the new location of TranslationContext and TranslationScenario.
